Feat: 질문 생성에 유저 정보 추가

+++ backend/src/main/java/com/ssafy/ssap/repository/UserRepository.java
@@ -1,17 +1,16 @@
 package com.ssafy.ssap.repository;
 
-import java.util.Optional;
-
+import com.ssafy.ssap.domain.user.User;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ssafy.ssap.domain.user.User;
+import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-	@EntityGraph(attributePaths = "authorities")
-	Optional<User> findOneWithAuthoritiesByEmail(String email);
+    @EntityGraph(attributePaths = "authorities")
+    Optional<User> findOneWithAuthoritiesByEmail(String email);
 
-	User findByEmail(String email);
+    User findByEmail(String email);
 
 }
